requirements.txt
scipy==1.10.1
gensim==3.7.0
jsonrpclib
nltk
numpy
redis
requests
sbd
stanford-corenlp
pexpect
py-rouge
